+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [Unreleased]
+### Changed
+- The exporter drops comments found in the `ipsec.conf`
+
 ## [0.3](https://github.com/dennisstritzke/ipsec_exporter/releases/tag/v0.3) - 2019-02-12
 ### Added
 - This Changelog.